refactor: group and extract cloud integration components to files
